requirements.txt
PyYAML~=6.0
Django~=3.2.9
pandas~=1.3.4
requests~=2.26.0
beautifulsoup4~=4.10.0